requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
APScheduler==3.10.4
python-dateutil==2.8.2
python-dotenv==1.0.0
marshmallow==3.20.1
